move Adapter class to src/
