+++ src/Adapter.js
